# javaChallenge/docSwagger.yml
swagger: "2.0"
info:
  description: "API desenvolvida para o teste de seleção para vaga de Java na empresa CD2. <br> Link do desafio: https://github.com/qualeloii/JavaTest"
  version: "1.0.0"
  title: "Siga Bem"

host: "aw-javachalleng.herokuapp.com"
tags:
- name: "Encomendas"
  description: ""
schemes:
- "http"
paths:
  /Encomendas:
    post:
      tags:
      - "Encomendas"
      summary: "Calcular frete"
      description: ""
      operationId: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Corpo da requisição"
        required: true
        schema:
          $ref: "#/definitions/requestDTO"
      responses:
        "201":
          description: "Created"
        "405":
          description: "Bad Request"
        "500":
          description: "Server Error"
    get:
      tags:
      - "Encomendas"
      summary: "Listar todos os fretes calculados"
      description: ""
      operationId: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "OK"
        "500":
          description: "Server Error"
definitions:
  requestDTO:
    type: "object"
    required:
    - "name"
    - "peso"
    - "cepOrigem"
    - "cepDestino"
    properties:
      name:
        type: "string"
        example: "Maria"
      peso:
        type: "number"
        multipleOf: 0.1
      cepOrigem:
        type: "string"
        example: "08090284"
      cepDestino:
        type: "string"
        example: "08090284"
  responseDTO:
    type: "object"
    properties:
      vlTotalFrete:
        type: "number"
        multipleOf: 0.1
      dataPrevistaEntrega:
        type: "string"
        example: "06/02/2022"
      cepOrigem:
        type: "string"
        example: "08090284"
      cepDestino:
        type: "string"
        example: "08090284"